Refactor: UI / Update css of nav components #605
