--- client/src/components/UI/IndexNav.module.css
@@ -1,17 +1,0 @@
-.container {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-top: var(--size-6);
-    padding: var(--size-10);
-    text-align: center;
-    font-size: var(--size-3);
-    font-weight: 600;
-    background-color: var(--gray-1);
-    border-radius: var(--size-6);
-}
-
-.container button {
-    padding: var(--size-6);
-}
