--- client/src/components/UI/StepNav.module.css
@@ -1,38 +1,0 @@
-.nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  text-align: right;
-}
-
-.nav button:only-child {
-  margin-left: auto;
-}
-
-.idx {
-  font-size: var(--size-4);
-}
-
-.prev,
-.next {
-  width: 4rem;
-  height: 4rem;
-  border-radius: 0.8rem;
-  font-size: var(--size-5);
-  font-weight: 600;
-}
-
-.prev {
-  text-align: left;
-  padding-left: 1rem;
-}
-
-.next {
-  background-color: var(--primary);
-  color: #fff;
-}
-
-.hide {
-  visibility: hidden;
-}
