+++ client/src/css/layout.css
@@ -14,6 +14,7 @@
   --round-xl: 1.75rem;
   --round-2lg: 1.5rem;
   --round-lg: 1rem;
+  --round-ml: 0.75rem;
   --round-md: 0.5rem;
   --round-sm: 0.375rem;
   --round-xs: 0.25rem;
@@ -127,6 +128,10 @@
 
 .mt-0\.5 {
   margin-top: 0.5rem;
+}
+
+.mt-1 {
+  margin-top: 1rem;
 }
 
 .mt-1\.75 {
@@ -397,6 +402,10 @@
 
 .round-lg {
   border-radius: var(--round-lg);
+}
+
+.round-ml {
+  border-radius: var(--round-ml);
 }
 
 .round-md {
