+++ client/src/screens/InitialSetting.tsx
@@ -15,7 +15,7 @@ import TransactionForm from '../components/Budget/Transaction/TransactionForm';
 import TransactionList from '../components/Budget/Transaction/TransactionList';
 import TransactionNav from '../components/Budget/Transaction/TransactionNav';
 import EmojiOverlay from '../components/UI/overlay/EmojiOverlay';
-import StepNav from '../components/UI/StepNav';
+import StepNav from '../components/UI/nav/StepNav';
 import DefaultStatus from '../components/User/Default/DefaultStatus';
 import { useAppSelector } from '../hooks/useRedux';
 import { assetActions } from '../store/asset';
